adminserver: fix non-deterministic spurious test failure (#609)
`AdminServer.Start()` almost always returns an error when the server is closed since the socket is closed while the HTTP server is in `accept`. This change eliminates a spurious test failure that seems to mostly occur in CI when the go routine running Start asserts after the test has passed.

fbshipit-source-id: 7715448