proc: fix non-determinism in tests
Introduces proc/proc-test which contains a main that produces specific
output and exits cleanly in response to signals.

Provide mechanism for creating a ManagedProc with a specific logger.
This is used in tests to examine proc-test output.

Rewrite proc tests to run proc-test and await output indicating that
it is ready before attempting to signal the process.

fbshipit-source-id: ef8d961